Refactors project and fixes proptype errors
