+++ lib/filter-tile-stars/five-stars.js
