+++ lib/filter-tile-stars/font-awesome-animation.css
