+++ lib/filter-tile-stars/index.js
