+++ lib/filter-tile-stars/info-bubble.js
