+++ lib/filter-tile-yesno/description.js
@@ -8,19 +8,19 @@ export default class Description extends Component {
     this.handleYesClick = this.handleYesClick.bind(this);
   }
   handleYesClick () {
-    const { description, yesFilter, noFilter, showAddMessage } = this.props;
+    const { description, yesFilter, showAddMessage } = this.props;
     yesFilter(description.bigWord);
     showAddMessage();
   }
   render () {
-    const { description, yesFilter, noFilter, showAddMessage } = this.props;
+    const { description, noFilter } = this.props;
     return (
       <div className='descriptionTextContainer'>
           <div className='description'>{description.part1}</div>
           <div className='bigWord'>{description.bigWord + ' ?'}</div>
           <div className='filterButtonContainer'>
-          <div className='filterButton' onClick={this.handleYesClick}>YES</div>
-          <div className='buttonSpacer'/>
+            <div className='filterButton' onClick={this.handleYesClick}>YES</div>
+            <div className='buttonSpacer'/>
             <div className='filterButton' onClick={() => noFilter(description.bigWord)}>NO</div>
           </div>
       </div>
