+++ lib/filter-tile-yesno/index.js
@@ -43,5 +43,7 @@ FilterTile.propTypes = {
   description: PropTypes.object,
   color: PropTypes.string,
   yesFilter: PropTypes.func,
-  noFilter: PropTypes.func
+  noFilter: PropTypes.func,
+  filterVisible: PropTypes.bool,
+  showAddMessage: PropTypes.func
 };
