+++ lib/filter-tile-yesno/infoText.js
