+++ src/actionCreators/filter.js
@@ -1,4 +1,4 @@
-import { YES_FILTER, NO_FILTER } from '../action-types.js';
+import { YES_FILTER, NO_FILTER } from '../actionTypes.js';
 
 export const yesFilter = (tagName) => { return {type: YES_FILTER, tagName}; };
 export const noFilter = (tagName) => { return {type: NO_FILTER, tagName}; };
