+++ src/actionCreators/tags.js
@@ -1,3 +1,3 @@
-import { REMOVE_TAG } from '../action-types.js';
+import { REMOVE_TAG } from '../actionTypes.js';
 
 export const removeTag = (tagName) => { return {type: REMOVE_TAG, tagName}; };
