+++ src/actionTypes.js
