+++ src/containers/home.js
@@ -1,6 +1,10 @@
 import React, { PropTypes, Component } from 'react';
 import { connect } from 'react-redux';
+
+// components
 import Home from '../components/home/';
+
+// actions
 import * as HomeActions from '../actionCreators/home.js';
 
 class HomeContainer extends Component {
@@ -14,7 +18,8 @@ class HomeContainer extends Component {
 }
 
 HomeContainer.propTypes = {
-  addMessageVisible: PropTypes.string,
+  // actions
+  addMessageVisible: PropTypes.bool,
   hideAddMessage: PropTypes.func
 };
