+++ src/containers/tags.js
@@ -1,6 +1,10 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
+
+// components
 import Tags from 'tags';
+
+// actions
 import * as TagsActions from '../actionCreators/tags.js';
 
 class TagsContainer extends Component {
@@ -15,14 +19,17 @@ class TagsContainer extends Component {
 }
 
 TagsContainer.propTypes = {
+  // actions
   removeTag: PropTypes.func,
+
+  // store
   tags: PropTypes.array
 };
 
 function mapStateToProps (state) {
-  return ({
+  return {
     tags: state.filter.tags
-  });
+  };
 }
 
 export default connect(mapStateToProps, TagsActions)(TagsContainer);
