+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import configureStore from './store/configure-store.js';
-import Root from './containers/root/index.js';
+import Root from './containers/index.js';
 
 require('./normalise.css');
